firmware: move clock-associated gpio functions to clkin.x and rename to clock_io.h|c
