+++ firmware/hackrf_usb/hackrf_usb.c
@@ -28,8 +28,8 @@
 #include <libopencm3/cm3/nvic.h>
 #include <libopencm3/lpc43xx/ipc.h>
 
-#include <clkin.h>
 #include <clock_gen.h>
+#include <clock_io.h>
 #include <cpu_clock.h>
 #include <da7219.h>
 #include <delay.h>
